Move models from db to its folder
